Add macros custom_filter(s) to resource classes
