+++ lib/jsonapi/utils.rb
@@ -3,6 +3,9 @@ require 'jsonapi/utils/version'
 require 'jsonapi/utils/exceptions'
 require 'jsonapi/utils/request'
 require 'jsonapi/utils/response'
+require 'jsonapi/utils/support/filter/custom'
+
+JSONAPI::Resource.extend JSONAPI::Utils::Support::Filter::Custom
 
 module JSONAPI
   module Utils
